Block spammers on deletion
