+++ test/fixtures/posts.yml
@@ -577,6 +577,23 @@ deleted_spam_question:
   deleted_at: 2019-01-01T00:00:00.000000Z
   deleted_by: moderator
 
+deleted_high_rep_spam:
+  post_type: question
+  title: Best spam in the neighborhood! Call 555-55-55
+  body: Find the best spam in your neighborhood on Spam Central
+  body_markdown: <p>Find the best spam in your neighborhood on <a href="https://example.com/spam">Spam Central</a></p>
+  tags_cache:
+    - discussion
+  tags:
+    - discussion
+  user: high_rep_spammer
+  community: sample
+  category: main
+  license: cc_by_sa
+  deleted: true
+  deleted_at: 2019-01-01T00:00:00.000000Z
+  deleted_by: moderator
+
 without_new_thread_followers:
   post_type: question
   title: This post does not have any new thread followers
