+++ test/jobs/delete_user_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class DeleteUserJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
